requirements.txt
Django==1.6.2
South==0.8.4
dj-static==0.0.5
django-allauth==0.15.0
django-autoslug==1.7.2
django-taggit==0.11.2
flake8==2.1.0
ipython==1.2.0
mccabe==0.2.1
oauthlib==0.6.1
pep8==1.4.6
pyflakes==0.7.3
pystache==0.5.3
python3-openid==3.0.3
requests==2.2.1
requests-oauthlib==0.4.0
static==1.0.2
